Drop use of qt5_use_modules for Qt 5.11 compatibility
As described in the qtbase 5.11 changelog[1], compatibility with versions
of CMake < 3.1 has been removed. In particular, the qt5_use_modules macro
is no longer available, requiring us to instead use target_link_libraries.

Alongside the macro changeover, we have also needed to add a new call to
find the Qt5MultimediaWidgets package which is used by demodatv, as well
as uncomment the finding of Qt5OpenGL.

[1]: http://code.qt.io/cgit/qt/qtbase.git/tree/dist/changes-5.11.0/?h=v5.11.0

Signed-off-by: Jason Gerecke <killertofu@gmail.com>