hackrfinput - Change LO ppm to adjust the hardware clocks.
LO ppm now adjusts the primary clock in the hackrf one instead of offsetting the freq in software. The offset adjusts all clocks except the CPU which currently uses it's own crystal.
There is room for improvement but I tried to keep it simple and keep the user interface the same for now.