Propose IssueTemplate
I would like to contribute to this project in the form of proposing an Issue Template. Reading the Wiki, I saw a section about how this project is fun when adding new features, and not so much when there are issues that turn out to not be reproducible, or worse, an issue with the hardware setup.

Hoping that this issue template can be added to the project and listed under the "Insights" tab of the community page.

Thank you for all that you have done in this project, I have learned so much!